requirements.txt
selenium
Pillow
easyocr
numpy
beautifulsoup4
pynput
torch
torchvision
